+++ tests/test_traits.py
@@ -1,3 +1,9 @@
+from pytest import raises
+
+
+from mio.errors import AttributeError
+
+
 def test_basic_trait(mio, capsys):
     mio.eval("""
         TGreeting = Object clone do (
@@ -52,7 +58,36 @@ def test_delTrait(mio):
 
     assert mio.eval("World hasTrait(TGreetable)").value is True
     mio.eval("World delTrait(TGreetable)")
+    assert mio.eval("World behaviors") == []
     assert mio.eval("World hasTrait(TGreetable)").value is False
+
+
+def test_delTrait2(mio, capsys):
+    mio.eval("""
+        TGreetable = Object clone do (
+            hello = method(
+                "Hello World!" println
+            )
+        )
+
+        World = Object clone do (
+            uses(TGreetable)
+        )
+    """)
+
+    assert mio.eval("World hasTrait(TGreetable)").value is True
+    assert mio.eval("World behaviors") == ["hello"]
+
+    assert mio.eval("World hello") == "Hello World!"
+    out, err = capsys.readouterr()
+    assert out == "Hello World!\n"
+
+    mio.eval("World delTrait(TGreetable)")
+    assert mio.eval("World behaviors") == []
+    assert mio.eval("World hasTrait(TGreetable)").value is False
+
+    with raises(AttributeError):
+        mio.eval("World hello", reraise=True)
 
 
 def test_addTrait(mio):
@@ -95,3 +130,28 @@ def test_behaviors(mio, capsys):
     out, err = capsys.readouterr()
     assert out == "Hello World!\n"
     assert mio.eval("World behaviors") == ["hello"]
+
+
+def test_del_behavior(mio, capsys):
+    mio.eval("""
+        TGreetable = Object clone do (
+            hello = method(
+                "Hello World!" println
+            )
+        )
+
+        World = Object clone do (
+            uses(TGreetable)
+        )
+    """)
+
+    assert mio.eval("World hello") == "Hello World!"
+    out, err = capsys.readouterr()
+    assert out == "Hello World!\n"
+    assert mio.eval("World behaviors") == ["hello"]
+
+    mio.eval("World del(\"hello\")")
+
+    with raises(AttributeError):
+        mio.eval("World hello", reraise=True)
+    assert mio.eval("World behaviors") == []